% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_agenda_slide.R
\name{create_agenda_slide}
\alias{create_agenda_slide}
\title{Create an Agenda Slide}
\usage{
create_agenda_slide(
  x,
  slide_title,
  slide_body = officer::unordered_list(level_list = c(1, 2, 3), str_list = c("Level1",
    "Level2", "Level1"), style = list(fp_text(color = "#C32900", font.size = 0),
    fp_text(color = "#006699", font.size = 0), fp_text(color = "#f2af00", font.size =
    0))),
  section_title = "Agenda",
  footer_text = "Client Engagement",
  layout_name = "Agenda",
  master_name = "EVA - Standard"
)
}
\arguments{
\item{x}{a PowerPoint presentation object.}

\item{slide_title}{Title of the slide.}

\item{slide_body}{Content of the slide, created using the officer::unordered_list function. Default list has 3 levels and is formatted with three different colors (red, blue, and orange).}

\item{section_title}{Title of the section.}

\item{footer_text}{Text to be displayed in the footer.}

\item{layout_name}{Name of the layout used for the slide. Default is "Agenda".}

\item{master_name}{Name of the master layout used for the slide. Default is "EVA - Standard".}
}
\value{
A PowerPoint slide object with the specified title, content, section title, and footer text.
}
\description{
This function creates an agenda slide in a PowerPoint presentation using the officer package in R.
}
\examples{
\dontrun{
ppt <- read_pptx()
create_agenda_slide(ppt, "Agenda", "Welcome", section_title = "Introduction")
}
}
