% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_column_slide.R
\name{create_column_slide}
\alias{create_column_slide}
\title{Create Column Slide}
\usage{
create_column_slide(
  x,
  slide_title = "Default Title for Columns",
  slide_caption = "",
  column_data = NULL,
  slide_section = "",
  footer_text = "",
  layout_name = "Two Column",
  master_name = "EVA - Standard"
)
}
\arguments{
\item{x}{The PowerPoint document to add the slide to.}

\item{slide_title}{The title to be added to the slide. Default is "Default Title for Columns".}

\item{slide_caption}{The caption to be added to the slide. Default is "".}

\item{column_data}{The data used to fill in the corresponding columns a list including title + body}

\item{slide_section}{The section to be added to the slide.}

\item{footer_text}{The footer text of the slide}

\item{layout_name}{The name of the slide layout to be used. Default is "Two Column".}

\item{master_name}{The name of the slide master to be used. Default is "EVA - Standard".}
}
\value{
A PowerPoint slide with the selected layout and added title, table, caption, body text and section.
}
\description{
This function creates a slide in a PowerPoint template containing a dynamic number of columns.
It adds a slide with the selected layout and adds a title, table, caption,
body text and section to the slide.
}
\examples{
\dontrun{
doc\%>\%create_column_slide(
x = .,
column_data = list(
list('title' = "Title of Column 1", 'body' = "Body of Column 1"),
list('title' = "Title of Column 2", 'body' = "Body of Column 2"),
list('title' = "Title of Column 3", 'body' = "Body of Column 3"),
list('title' = "Title of Column 4", 'body' = "Body of Column 4")
),
layout_name = "Four Column"
)
}
}
