% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_decision_point_slide.R
\name{create_decision_point_slide}
\alias{create_decision_point_slide}
\title{Create Decision Point Slide}
\usage{
create_decision_point_slide(
  x,
  dp_header = "Primary decision point text",
  actions_body = officer::unordered_list(level_list = c(1, 2, 3), str_list = c("Level1",
    "Level2", "Level1"), style = list(fp_text(color = "#000000", font.size = 0),
    fp_text(color = "#000000", font.size = 0), fp_text(color = "#000000", font.size =
    0))),
  decision_body = officer::unordered_list(level_list = c(1, 2, 3), str_list = c("Level1",
    "Level2", "Level1"), style = list(fp_text(color = "#000000", font.size = 0),
    fp_text(color = "#000000", font.size = 0), fp_text(color = "#000000", font.size =
    0))),
  section_text = "Section",
  date_text = Sys.Date(),
  attendees_list = officer::unordered_list(level_list = c(1, 1), str_list = c("Level1",
    "Level2"), style = list(fp_text(color = "#000000", font.size = 0), fp_text(color =
    "#000000", font.size = 0))),
  footer_text = "",
  layout_name = "Decision Point",
  master_name = "EVA - Standard"
)
}
\arguments{
\item{x}{The PowerPoint document to add the slide to.}

\item{dp_header}{The main header text for the decision point slide.}

\item{actions_body}{The body text for the actions section of the slide. Default is an empty unordered list.}

\item{decision_body}{The body text for the decision section of the slide. Default is an empty unordered list.}

\item{section_text}{The section text for the slide. Default is "Section".}

\item{date_text}{The date to be displayed on the slide. Default is the current date.}

\item{attendees_list}{The list of attendees to be displayed on the slide. Default is an empty unordered list.}

\item{footer_text}{The footer text to be displayed on the slide. Default is an empty string.}

\item{layout_name}{The name of the slide layout to be used. Default is "Decision Point".}

\item{master_name}{The name of the slide master to be used. Default is "EVA - Standard".}
}
\value{
A PowerPoint slide with the selected layout and added header, actions, decision, section, date, attendees, and footer.
}
\description{
This function creates a decision point slide in a PowerPoint template. It adds a slide with the selected layout and adds a main header, actions, decision, section, date, attendees, and footer to the slide.
}
\examples{
\dontrun{
create_decision_point_slide(doc,
"Main Decision",
actions_body,
decision_body,
"Section1",
"2023-02-22",
attendees_list,
"Footer")
}
}
