% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_graphic_slide.R
\name{create_graphic_slide}
\alias{create_graphic_slide}
\title{Create Graphic Slide}
\usage{
create_graphic_slide(
  x,
  slide_title = "Default Title for Graphic",
  slide_caption = "",
  slide_graphic,
  slide_body = "Some text",
  slide_section = "",
  footer_text = "",
  layout_name = "Graphic Full",
  master_name = "EVA - Standard"
)
}
\arguments{
\item{x}{The PowerPoint document to add the slide to.}

\item{slide_title}{The title to be added to the slide. Default is "Default Title for Table".}

\item{slide_caption}{The caption to be added to the slide. Default is "".}

\item{slide_graphic}{The graphic to be added to the slide.}

\item{slide_body}{The body text to be added to the slide. Default is "Some text".}

\item{slide_section}{The section to be added to the slide.}

\item{footer_text}{The footer text of the slide}

\item{layout_name}{The name of the slide layout to be used. Default is "graphic Full".}

\item{master_name}{The name of the slide master to be used. Default is "EVA - Standard".}
}
\value{
A PowerPoint slide with the selected layout and added title, table, caption, body text and section.
}
\description{
This function creates a slide in a PowerPoint template containing a graphic.
It adds a slide with the selected layout and adds a title, table, caption,
body text and section to the slide.
}
\examples{
\dontrun{
create_table_slide(x,
"Table Title",
"Table Caption",
my_graphic,
"Some text",
"Section A",
"Graphic Full",
"EVA - Standard",
F)
}
}
