% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_point_slide.R
\name{create_point_slide}
\alias{create_point_slide}
\title{Create Point Slide}
\usage{
create_point_slide(
  x,
  slide_title = "Default Title for Point Slide",
  slide_subtitle = "Default Subtitle for Point Slide",
  slide_body = "Body Text",
  footer_text = "",
  layout_name = "Point - Detail Bullet",
  master_name = "EVA - Standard"
)
}
\arguments{
\item{x}{The PowerPoint document to add the slide to.}

\item{slide_title}{The title to be added to the slide.}

\item{slide_subtitle}{The subtitle to be added to the slide.}

\item{slide_body}{The body text or contents to be added to the slide.}

\item{footer_text}{The footer text of the slide}

\item{layout_name}{The name of the slide layout to be used. Default is "Title - Black".}

\item{master_name}{The name of the slide master to be used. Default is "EVA - Standard".}
}
\value{
A PowerPoint slide with the selected layout and added title, subtitle and picture.
}
\description{
This function creates a point slide in a PowerPoint template. It adds
a slide with the selected layout and adds a title, subtitle and picture
to the slide.
}
\examples{
\dontrun{
create_point_slide(x, "Title", "Subtitle", "Body Text", "Point - Detail Bullet", "EVA - Standard")
}
}
