% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_problem_statement_slide.R
\name{create_problem_statement_slide}
\alias{create_problem_statement_slide}
\title{Create Problem Statement Slide}
\usage{
create_problem_statement_slide(
  x,
  slide_problem_statement = "Insert Problem Statement Here",
  slide_subtitle = "Caption for the problem statement",
  slide_body = "Body Text",
  layout_name = "Problem Statement",
  master_name = "EVA - Standard"
)
}
\arguments{
\item{x}{The PowerPoint document to add the slide to.}

\item{slide_problem_statement}{The title to be added to the slide.}

\item{slide_subtitle}{The subtitle to be added to the slide.}

\item{slide_body}{The body text or contents to be added to the slide.}

\item{layout_name}{The name of the slide layout to be used. Default is "Problem Statement".}

\item{master_name}{The name of the slide master to be used. Default is "EVA - Standard".}
}
\value{
A PowerPoint slide with the selected layout and added title, subtitle
and picture.
}
\description{
This function creates a problem statement slide in a PowerPoint template. It adds
a slide with the selected layout and adds a title, subtitle and body text for
additional details to the slide.
}
\examples{
\dontrun{
create_problem_state_slide(x, "Title", "Subtitle", "Body Text",
"Problem Statement", "EVA - Standard")
}
}
