% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_section_slide.R
\name{create_section_slide}
\alias{create_section_slide}
\title{Create Section Slide}
\usage{
create_section_slide(
  x,
  slide_title,
  slide_subtitle,
  layout_name = "Section - Black",
  master_name = "EVA - Standard",
  logo = NULL
)
}
\arguments{
\item{x}{The PowerPoint document to add the slide to.}

\item{slide_title}{The title to be added to the slide.}

\item{slide_subtitle}{The subtitle to be added to the slide.}

\item{layout_name}{The name of the slide layout to be used. Default is "Section - Black".}

\item{master_name}{The name of the slide master to be used. Default is "EVA - Standard".}

\item{logo}{The location of a logo file for a client presentation. If NULL, ignore}
}
\value{
A PowerPoint slide with the selected layout and added title, subtitle and picture.
}
\description{
This function creates a title slide in a PowerPoint template. It adds a slide with the selected layout and adds a title and subtitle to the slide.
}
\examples{
\dontrun{
create_title_slide(doc,
"Title",
"Subtitle",
"Section - Blue",
"EVA - Standard")
}
}
