% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_slide.R
\name{create_table_slide}
\alias{create_table_slide}
\title{Create Table Slide}
\usage{
create_table_slide(
  x,
  slide_title = "Default Title for Table",
  slide_caption = "",
  slide_table,
  slide_body = "Some text",
  slide_section = "",
  footer_text = "",
  layout_name = "Table Full",
  master_name = "EVA - Standard",
  autofit = F
)
}
\arguments{
\item{x}{The PowerPoint document to add the slide to.}

\item{slide_title}{The title to be added to the slide. Default is "Default Title for Table".}

\item{slide_caption}{The caption to be added to the slide. Default is "".}

\item{slide_table}{The table to be added to the slide.}

\item{slide_body}{The body text to be added to the slide. Default is "Some text".}

\item{slide_section}{The section to be added to the slide.}

\item{footer_text}{The footer text of the slide}

\item{layout_name}{The name of the slide layout to be used. Default is "Table Full".}

\item{master_name}{The name of the slide master to be used. Default is "EVA - Standard".}

\item{autofit}{A boolean to specify if the table should be autofitted. Default is F.}
}
\value{
A PowerPoint slide with the selected layout and added title, table, caption, body text and section.
}
\description{
This function creates a slide in a PowerPoint template containing a table. It adds a slide with the selected layout and adds a title, table, caption, body text and section to the slide.
}
\examples{
\dontrun{
create_table_slide(x,
"Table Title",
"Table Caption",
my_table,
"Some text",
"Section A",
"Table Full",
"EVA - Standard",
F)
}
}
