% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph_disclaimer.R
\name{ph_disclaimer}
\alias{ph_disclaimer}
\title{Add a disclaimer to a PowerPoint slide}
\usage{
ph_disclaimer(x, val = "PRIVATE AND STRICTLY CONFIDENTIAL")
}
\arguments{
\item{x}{A PowerPoint document.}

\item{val}{The text to use for the disclaimer. Default is "PRIVATE AND STRICTLY CONFIDENTIAL".}
}
\value{
A PowerPoint slide with a disclaimer text box.
}
\description{
This function adds a disclaimer text to a PowerPoint slide. By default, the disclaimer text is "PRIVATE AND STRICTLY CONFIDENTIAL", but it can be customized to a different text by providing a value for the "val" parameter.
}
\examples{
\dontrun{
ph_disclaimer(my_presentation, "This information is for internal use only.")
}
}
