% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph_slide_number.R
\name{ph_slide_number}
\alias{ph_slide_number}
\title{Add slide numbers to PowerPoint slides}
\usage{
ph_slide_number(x)
}
\arguments{
\item{x}{A PowerPoint document.}
}
\value{
A PowerPoint slide with a slide number.
}
\description{
This function adds slide numbers to a PowerPoint slide. It retrieves the maximum slide ID using pptx_summary and adds it to the slide as a text box.
}
\examples{
\dontrun{
ph_slide_number(my_presentation)
}
}
