% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ppt_template.R
\name{read_ppt_template}
\alias{read_ppt_template}
\title{Read PowerPoint Template}
\usage{
read_ppt_template(
  template_path = here::here("inst/extdata", "default-eva-template.potx")
)
}
\arguments{
\item{template_path}{The complete location of the powerpoint template.}
}
\value{
A pptx object for further modifications.
}
\description{
This function reads in a PowerPoint template file and returns a pptx object to be used for further modifications.
}
\examples{
\dontrun{
doc <- read_ppt_template("default-eva-template.potx")
}

}
