% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ppt_template.R
\name{read_ppt_template}
\alias{read_ppt_template}
\title{Read PowerPoint Template}
\usage{
read_ppt_template(
  folder_location = "data",
  document_name = "default-eva-template.potx"
)
}
\arguments{
\item{folder_location}{The folder that hosts the powerpoint template}

\item{document_name}{The name of the PowerPoint template file to be read in.}
}
\value{
A pptx object for further modifications.
}
\description{
This function reads in a PowerPoint template file and returns a pptx object to be used for further modifications.
}
\examples{
\dontrun{
doc <- read_ppt_template("default-eva-template.potx")
}

}
